+++ labelme/_utils/shape.py
@@ -5,12 +5,15 @@ from __future__ import annotations
 
 import math
 import uuid
+from typing import NotRequired
 from typing import TypedDict
 
 import numpy as np
 import PIL.Image
 import PIL.ImageDraw
 from numpy.typing import NDArray
+
+from labelme._line_profile import LineProfile
 
 
 class ShapeDict(TypedDict):
@@ -22,6 +25,8 @@ class ShapeDict(TypedDict):
     group_id: int | None
     mask: NDArray[np.bool_] | None
     other_data: dict
+    line_profile: NotRequired[LineProfile | None]
+    line_profile_error: NotRequired[str | None]
 
 
 def shape_to_mask(